requirements.txt
crewai==0.117.0
langchain-community
duckduckgo-search
langchain-google-genai
python-dotenv